feat: phil2 initial commit - rewrite phil to use slash commands
Discord intends to limit the ability of bots to read messages except where they have a privileged intent. This bot doesn't strictly need to be abel to read message contents so instead of applying for the intent need to rewrite to use slash commands.

This is a good opportunity to start rewriting in a more modular fashion so that the dreams of Silicon Dan can eventually be realised